chore(ci): update GHA workflows to more closely match Pillow's
This fixes issues with outdated action versions and runner images in
our CI workflows, and integrates cibuildwheel in a manner similar to
Pillow. This increases the number of platforms and python versions for
which we can build wheels and reduces the overall build time.

winbuild/build_prepare.py has likewise been updated to more
closely resemble the version of that file in the Pillow repository.